Bluetooth-OPP: Use long instead of int
This patch fixes issue of problems in sending/receiving files
which cannot be accomodated in int type of Java. Instead of using
int, long data type is used.

+++ src/com/android/bluetooth/opp/BluetoothOppObexServerSession.java
@@ -109,7 +109,7 @@ public class BluetoothOppObexServerSession extends ServerRequestHandler implemen
 
     boolean mTransferInProgress = false;
 
-    private int position;
+    private long  position;
 
     public BluetoothOppObexServerSession(Context context, ObexTransport transport) {
         mContext = context;
